kas umum revisi

+++ application/controllers/KasKeluar.php
@@ -75,14 +75,42 @@ class KasKeluar extends CI_Controller
     }
     public function hapus($kode)
     {
-        $b = $this->db->query("SELECT * FROM tb_kaskeluar WHERE kode_kas_keluar = '" . $kode . "'")->row_array();
-        $a = $this->db->query('SELECT * FROM tb_historikas ORDER BY id_histori_kas DESC LIMIT 1')->row_array();
-        $hasil = intval($a['saldo']) + intval($b['nominal']);
-        $this->db->where('kode_kas', $a['kode_kas']);
-        $this->db->update('tb_historikas', ['saldo' => $hasil]);
-        $this->db->where('kode_kas', $kode);
-        $this->db->delete('tb_historikas');
-        $this->M_KasKeluar->hapus($kode);
+
+        $b1 = $this->db->query("SELECT * FROM tb_historikas WHERE kode_kas = '" . $kode . "'")->row_array();
+        $querya = "SELECT * FROM tb_historikas";
+        $a = $this->db->query($querya)->result_array();
+        $j = $this->db->query($querya)->num_rows();
+        $i = 0;
+        foreach ($a as $data) {
+            $i++;
+            if ($data['kode_kas'] == $kode) {
+                $awal = $i;
+                $queryb = "SELECT * FROM tb_historikas LIMIT " . intval($awal) . "," . intval($j);
+                $b = $this->db->query($queryb)->result_array();
+                $c = $this->db->query($queryb)->num_rows();
+
+                if ($c >= 1) {
+                    foreach ($b as $data1) {
+                        $hasil = $data1['saldo'] + $b1['nominal'];
+                        $datah = [
+                            'saldo' => $hasil
+                        ];
+                        $this->db->where('kode_kas', $data1['kode_kas']);
+                        $this->db->update('tb_historikas', $datah);
+                        $this->db->where('kode_kas', $kode);
+                        $this->db->delete('tb_historikas');
+                        $this->db->where('kode_kas_keluar', $kode);
+                        $this->db->delete('tb_kaskeluar');
+                    }
+                } else {
+                    $this->db->where('kode_kas', $kode);
+                    $this->db->delete('tb_historikas');
+                    $this->db->where('kode_kas_keluar', $kode);
+                    $this->db->delete('tb_kaskeluar');
+                }
+            }
+        }
+
         $this->session->set_flashdata('message', '<div class="alert alert-success left-icon-alert" role="alert"> <strong>Sukses!</strong> Data Berhasil DiHapus</div>');
         redirect('kaskeluar');
     }
@@ -99,35 +127,109 @@ class KasKeluar extends CI_Controller
     }
     public function ubahdata()
     {
-        $kodekaskeluar = $this->input->post('kode');
+        $kode = $this->input->post('kode');
         $id = $this->session->userdata('tipeuser');
-        $saldo = $this->db->query("SELECT * FROM tb_historikas ORDER BY id_histori_kas DESC LIMIT 1")->row_array();
-        $hasil = intval($saldo['saldo']) - intval(preg_replace("/[^0-9]/", "", $this->input->post('nominal')));
-        $data = [
-            'tgltransaksi' => $this->input->post('tglTransaksi') . date('h:i:s'),
-            'keterangan' => $this->input->post('keterangan'),
-            'nominal' => preg_replace("/[^0-9]/", "", $this->input->post('nominal')),
-            'kode_kas_keluar' => $kodekaskeluar,
-            'id_user' => $id,
-        ];
-        // var_dump($data);
-        if (preg_replace("/[^0-9]/", "", $this->input->post('nominal')) > intval($saldo['saldo'])) {
+        $nominal = preg_replace("/[^0-9]/", "", $this->input->post('nominal'));
+        $b1 = $this->db->query("SELECT * FROM tb_historikas WHERE kode_kas = '" . $kode . "'")->row_array();
+        $dbet = $this->db->query("SELECT SUM(nominal) AS nominal FROM tb_historikas WHERE jenis ='kas masuk'")->row_array();
+        $kreddi = $this->db->query("SELECT SUM(nominal) AS nominal FROM tb_historikas WHERE jenis='kas keluar'")->row_array();
+        $saldo = $dbet['nominal'] - $kreddi['nominal'];
+        $querya = "SELECT * FROM tb_historikas";
+        $a = $this->db->query($querya)->result_array();
+        $j = $this->db->query($querya)->num_rows();
+        $i = 0;
+        if ($nominal > $saldo) {
             $this->session->set_flashdata('message', '<div class="alert alert-warning left-icon-alert" role="alert"> <strong>Warning! </strong>Nominal Terlalu Besar</div>');
-            redirect('kaskeluar/tambah');
+            redirect('kas-keluar-edt/' . $kode);
         } else {
-            $this->M_KasKeluar->ubah($data, $kodekaskeluar);
-            $this->db->where('kode_kas', $kodekaskeluar);
-            $this->db->delete('tb_historikas');
-            $dataHistori = [
-                'kode_kas' => $kodekaskeluar,
-                'jenis' => 'kas keluar',
-                'nominal' => preg_replace("/[^0-9]/", "", $this->input->post('nominal')),
-                'saldo' => $hasil,
-                'tgltransaksi' => $this->input->post('tglTransaksi') . date(' h:i:s'),
-            ];
+            foreach ($a as $data) {
+                $i++;
+                if ($data['kode_kas'] == $kode) {
+                    $awal = $i;
+                    $queryb = "SELECT * FROM tb_historikas LIMIT " . intval($awal) . "," . intval($j);
+                    $b = $this->db->query($queryb)->result_array();
+                    $c = $this->db->query($queryb)->num_rows();
 
-            $this->M_KasKeluar->tambahHisto($dataHistori);
+                    if ($c >= 1) {
+                        foreach ($b as $data1) {
+                            $saldodia = $b1['saldo'];
+                            $hasildia = $saldodia + $b1['nominal'] - $nominal;
+                            $hasil = $hasildia - $data1['nominal'];
+                            $datadia = [
+                                'nominal' => $nominal,
+                                'saldo' => $hasildia,
+                                'tgltransaksi' => $this->input->post('tglTransaksi') . date(' h:i:s')
+                            ];
+
+                            $datar = ['saldo' => $hasil];
+
+                            $data = [
+                                'tgltransaksi' => $this->input->post('tglTransaksi') . date(' h:i:s'),
+                                'keterangan' => $this->input->post('keterangan'),
+                                'nominal' => preg_replace("/[^0-9]/", "", $this->input->post('nominal')),
+                                'kode_kas_keluar' => $kode,
+                                'id_user' => $id,
+                                'status_jurnal' => '0'
+                            ];
+                            $this->M_KasKeluar->ubah($data, $kode);
+                            $this->db->where('kode_kas', $kode);
+                            $this->db->update('tb_historikas', $datadia);
+                            $this->db->where('kode_kas', $data1['kode_kas']);
+                            $this->db->update('tb_historikas', $datar);
+                        }
+                    } else {
+                        $hasildia = $saldo + $b1['nominal'] - $nominal;
+
+                        $datar = [
+                            'nominal' => $nominal,
+                            'saldo' => $hasildia,
+                            'tgltransaksi' => $this->input->post('tglTransaksi') . date(' h:i:s')
+                        ];
+                        $data = [
+                            'tgltransaksi' => $this->input->post('tglTransaksi') . date(' h:i:s'),
+                            'keterangan' => $this->input->post('keterangan'),
+                            'nominal' => preg_replace("/[^0-9]/", "", $this->input->post('nominal')),
+                            'kode_kas_keluar' => $kode,
+                            'id_user' => $id,
+                            'status_jurnal' => '0'
+                        ];
+                        $this->M_KasKeluar->ubah($data, $kode);
+                        $this->db->where('kode_kas', $kode);
+                        $this->db->update('tb_historikas', $datar);
+                    }
+                }
+            }
         }
+
+        // $kodekaskeluar = $this->input->post('kode');
+        // $id = $this->session->userdata('tipeuser');
+        // $saldo = $this->db->query("SELECT * FROM tb_historikas ORDER BY id_histori_kas DESC LIMIT 1")->row_array();
+        // $hasil = intval($saldo['saldo']) - intval(preg_replace("/[^0-9]/", "", $this->input->post('nominal')));
+        // $data = [
+        //     'tgltransaksi' => $this->input->post('tglTransaksi') . date('h:i:s'),
+        //     'keterangan' => $this->input->post('keterangan'),
+        //     'nominal' => preg_replace("/[^0-9]/", "", $this->input->post('nominal')),
+        //     'kode_kas_keluar' => $kodekaskeluar,
+        //     'id_user' => $id,
+        // ];
+        // // var_dump($data);
+        // if (preg_replace("/[^0-9]/", "", $this->input->post('nominal')) > intval($saldo['saldo'])) {
+        //     $this->session->set_flashdata('message', '<div class="alert alert-warning left-icon-alert" role="alert"> <strong>Warning! </strong>Nominal Terlalu Besar</div>');
+        //     redirect('kas-keluar-edt/' . $kodekaskeluar);
+        // } else {
+        //     $this->M_KasKeluar->ubah($data, $kodekaskeluar);
+        //     $this->db->where('kode_kas', $kodekaskeluar);
+        //     $this->db->delete('tb_historikas');
+        //     $dataHistori = [
+        //         'kode_kas' => $kodekaskeluar,
+        //         'jenis' => 'kas keluar',
+        //         'nominal' => preg_replace("/[^0-9]/", "", $this->input->post('nominal')),
+        //         'saldo' => $hasil,
+        //         'tgltransaksi' => $this->input->post('tglTransaksi') . date(' h:i:s'),
+        //     ];
+
+        //     $this->M_KasKeluar->tambahHisto($dataHistori);
+        // }
 
         $this->session->set_flashdata('message', '<div class="alert alert-success left-icon-alert" role="alert"> <strong>Sukses!</strong> Data Berhasil DiUbah</div>');
         redirect('kaskeluar');
